backend/pyproject.toml
[tool.poetry]
name = "youmail-backend"
version = "0.1.0"
description = "Mail Web Application backend with Gmail API, gRPC and FastAPI"
authors = ["YouMail"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.100.0"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
fastapi = "^0.109.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
python-dotenv = "^1.0.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
